Split logging sinks into separate files
